vision-service/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
torch>=2.1
timm>=1.0
segmentation-models-pytorch>=0.3
transformers>=4.41
qdrant-client>=1.11
Pillow>=10.0
numpy>=1.26
python-multipart>=0.0.9
